Remove custom types folder, just reduce it to the transaction options
Make args types consistent
